Change ts file structure JC-633
