+++ apps/vanilla/src/pages/task2/index.html
@@ -50,7 +50,6 @@
 			<button class="main__button">Roll the dice</button>
 		</section>
 	</main>
-	<script type="module" src="./main.ts"></script>
-	<script type="module" src="./layout.ts"></script>
+	<script type="module" src="./ts/main.ts"></script>
 </body>
 </html>
