+++ apps/vanilla/src/pages/task2/ts/constants.ts
@@ -1,0 +1,5 @@
+/** This is a description of the foo function. */
+export const PLAYERS_COUNT = 2;
+
+/** This is a description of the foo function. */
+export const SIDES_COUNT = 6;
