+++ apps/vanilla/src/pages/task2/ts/types/playerTurnResult.ts
@@ -1,0 +1,9 @@
+/** This is a description of the foo function. */
+export type PlayerTurnResult = {
+
+	/** This is a description of the foo function. */
+	readonly playerIndex: number;
+
+	/** This is a description of the foo function. */
+	readonly diceResult: number;
+};
