+++ apps/vanilla/src/pages/task2/ts/types/subscriber.ts
@@ -1,0 +1,6 @@
+/** This is a description of the foo function. */
+export type Subscriber<T> = {
+
+	/** This is a description of the foo function. */
+	readonly update: (message: T) => void;
+};
